+++ server/schemas/sleep.py
@@ -4,5 +4,5 @@ class TrackerRequest(BaseModel):
     wakeup: str
     bedtime: str
     awakenings: int
-    timeInBed: int
+    timeInBed: float
     isGoodSleep: bool
